Removed Switch statement, added more output
I removed the Switch statement that asked if you wanted to Protect your Privacy by changing specific registry key values and creating specific ones as well as disabling scheduled tasks.

The reason is that even if you chose to revert changes, it still asked you if you wanted to Protect your privacy. The problem with that is when you choose to Revert Changes, some key values and the scheduled tasks are re-enabled, but then afterwards when it asks if you want to protect your privacy it would revert the changes that the Revert-Changes function did, so it was not only redundant but self-defeating.

This is now fixed.