New Stop-EdgePDF function, new registry keys, merged switch statements
#Line 289
Added } at the end of If statement since it was missing.

#Line 358
I have changed the code to not automatically stop Edge from being the PDF viewer. It is now a Switch statement nested in another switch statement, so that you can choose whether or not to stop Edge from being the default PDF Viewer.

#Line 344
I also removed all locations where the PSDrive 'HKCR' was created, and added it to the Debloat switch statement, so that it is only created once for the entire session instead of needing to create it multiple times.

#Line 374
At the end of the code, where it asks if you'd like to reboot your machine, it will unload the PSDrive 'HKCR', whether or not you choose 'Yes' or 'No'.

#Line 130
I added a new registry key to remove the People icon from the taskbar. However, this will ONLY remove it from the current user's profile on the same machine, so if you want to remove it from multiple user profiles on one machine you will need to either add this key to each profile when logged in with that account or run this script. 

#Line 296
I also added code to bring back the People icon if you wish in the Revert-Changes function.

#Line 92
I added more error handling to the code that creates 3 registry key directories in HKCU:\Software\Microsoft\Siuf\Rules\PeriodInNanoSeconds.

#Line 385
I moved the code at the end of the script telling you that it has finished and that it was going to exit to the inside of the switch statement where it asks if you'd like to reboot. The portion where it tells you it is finished and going to exit is underneath the "No" portion of the switch statement.