Disable/enable start menu suggestions, new function and switch statement
#Line 143
Added code in the function Protect-Privacy to change the registry key property SystemPaneSuggestionsEnabled to a value of 0 to disable suggestions on the start menu

#Line 279
Added code in the function Revert-Changes to change the registry key property SystemPaneSuggestionsEnabled to a value of 1 to enable suggestions on the start menu

#Line 295
Created new function, Enable-EdgePDF. This will contain the code from Revert-Changes to re-enable Edge to act as the default PDF viewer. I removed the code from the Revert-Changes function since it didn't make sense to keep it in there since the Stop-EdgePDF function was created to disable Edge as the Default PDF Viewer, and I wanted to make it follow a pattern.

#Line 421
Added switch statement to ask if you'd like to revert changes to allow Edge to be used as the default PDF viewer.